Move ReactSliderShadowNode to shared implementation
